requirements.txt
requests
python-dotenv==1.0.1
fastapi==0.111.0
Telethon==1.36.0
uvicorn==0.30.1
pydantic==2.8.2